% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildTable.R
\name{BuildTable}
\alias{BuildTable}
\title{Build Table}
\usage{
BuildTable(M, Pi, offest, model, alpha = 1, beta = 1)
}
\arguments{
\item{value}{text}
}
\description{
More description
}
\examples{


}
\keyword{text}
