% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elton-package.R
\docType{package}
\name{elton-package}
\alias{elton-package}
\alias{elton}
\title{Multispecies distribution modeling}
\description{
Eltonian niche concepts, cutting edge methods for species distribution modeling with food webs or association networks.
}
\details{
{\emph{And all this science I don't understand, it's just my job five days a week}

--Elton John, "Rocket Man"}
}
\references{
Staniczenko, P.P., Sivasubramaniam, P., Suttle, K.B. and Pearson, R.G., 2017. Linking macroecology and community ecology: refining predictions of species distributions using biotic interaction networks. Ecology letters, 20(6), pp.693-707.
}
\author{
Colin Carlson (\email{ccarlson@sesync.org})

Phillip Staniczenko (\email{pstaniczenko@sesync.org})
}
