% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{GetBinaryLinearFacInhResponse}
\alias{GetBinaryLinearFacInhResponse}
\title{GetBinaryLinearFacInhResponse}
\usage{
GetBinaryLinearFacInhResponse(netcontributingunits, maxunits, p)
}
\arguments{
\item{value}{text}
}
\description{
More description
}
\keyword{text}
