% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{GetBinaryNonLinearFacInhResponse}
\alias{GetBinaryNonLinearFacInhResponse}
\title{GetBinaryNonLinearFacInhResponse}
\usage{
GetBinaryNonLinearFacInhResponse(netcontributingunits, posEdges, negEdges,
  maxunits, p, a, b)
}
\arguments{
\item{value}{text}
}
\description{
More description
}
\keyword{text}
