% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{GetBinaryORFacInhResponse}
\alias{GetBinaryORFacInhResponse}
\title{GetBinaryORFacInhResponse}
\usage{
GetBinaryORFacInhResponse(netcontributingunits, maxunits, p)
}
\arguments{
\item{value}{text}
}
\description{
More description
}
\keyword{text}
