% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMarginals_gRain.R
\name{getMarginals_gRain}
\alias{getMarginals_gRain}
\title{Run gRain and get all the results}
\usage{
getMarginals_gRain(Table, file = "test.R")
}
\arguments{
\item{value}{text}
}
\description{
Pull dependent species
}
\examples{

}
\keyword{text}
