% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isDAG.R
\name{isDAG}
\alias{isDAG}
\title{Is it a DAG?}
\usage{
isDAG(myDAG)
}
\arguments{
\item{myDAG}{your input graph, in the form of a matrix with values -1, 0, or +1}
}
\description{
Check if a graph is directed and acyclic, converting the interaction matrix into a positive-only format that can be checked by igraph::is.dag
}
\keyword{acyclic}
\keyword{directed}
\keyword{graph}
