% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{posteriors}
\alias{posteriors}
\title{Pull posteriors from gRain results}
\usage{
posteriors(grain.results)
}
\arguments{
\item{value}{text}
}
\description{
Pull out a clean vector of posterior probabilities (useful for raster work)
}
\examples{

}
\keyword{text}
