% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{TicksResponse}
\alias{TicksResponse}
\title{TickResponse}
\usage{
TicksResponse(netcontributingunits, posEdges, negEdges, maxunits, p, offset, a,
  b)
}
\arguments{
\item{value}{text}
}
\description{
More description
}
\keyword{text}
